Part of https://github.com/zed-industries/zed/issues/10525

Move `mimalloc` under the feature flag, `cargo build --features mimalloc` could be used to restore the old behavior.


Release Notes:

- Switched back to system allocator, keeping mimalloc as a compilation option