@mrnugget 2024-05-02
I did the same benchmarks as described in https://github.com/zed-industries/zed/pull/7140#issuecomment-2059079602 again, this time paying attention to the values when the window is focused.

### Methodology:

- Run `cargo run --release <projectfolder>`
- Close every open file
- Only keep project panel open
- Quit process
- Run `cargo run --release <projectfolder>` again
- Open `Activity Monitor.app`
- Have both — `Zed` and `Activity Monitor` — visible
- Focus Zed! Note the memory consumption in `Activity Monitor` changing.
- Write down `Before loading`
- Open file
- Wait 5sec, write down `After loading`
- Close file
- Wait 5sec, write down `After closing`


### Results:

| What                          | Allocator  | Before loading | After loading | After closing | Comment                            |
| ----------------------------- | ---------- | -------------- | ------------- | ------------- | ---------------------------------- |
| 132MB `.tex` file             | `mimalloc` | 448 MB         | 1660 MB        | 1530 MB        | No tex extension installed         |
| 10KB `.zshrc` file            | `mimalloc` | 463 MB         | 472 MB        | 473 MB        | -                                  |
| 403KB `editor.rs` file in Zed | `mimalloc` | 469 MB         | 511 MB        | 504 MB        | Waited until rust-analyzer started |
| 132MB `.tex` file             | `default`  | 422 MB         | 991 MB        | 460 MB        | No tex extension installed         |
| 10KB `.zshrc` file            | `default`  | 433 MB         | 443 MB        | 446 MB        | -                                  |
| 403KB `editor.rs` file in Zed | `default`  | 435 MB         | 473 MB        | 460 MB        | Waited until rust-analyzer started |

### Frametimes:

(Following these instructions: https://github.com/zed-industries/zed/pull/7140#issuecomment-1918887685 Except that I didn't restart and didn't close all other processes, I have 128GB of memory)

![screenshot-2024-05-02-12 51 50@2x](https://github.com/zed-industries/zed/assets/1185253/a72ca2cf-c738-4790-99b1-dc19bba95628)


### Conclusion?

I think this time the results are pretty clear:

* mimalloc is very aggressive in allocating memory (it allocated 1.6GB when opening the text file vs. the 991 MB the default allocator allocated)
* mimalloc is very defensive in deallocating memory (it only goes back to 1.5GB when closing the file, vs. 460MB (near baseline) of default allocator)
* frametimes don't suggest that mimalloc is faster.

So I think we should switch back to the default allocator.