@SomeoneToIgnore 2024-05-02
Thank you for testing and verifying this.

Probably worth another set of patch releases, but let's at least merge it for now.